+++ app/src/main/java/com/example/actorsdatabaseapp/sql/data/model/ActorMovies.kt
@@ -1,3 +1,3 @@
 package com.example.actorsdatabaseapp.sql.data.model
 
-data class ActorMovies(val actorId: Int, val actorName: String, val movieRate: Int, val movieYear: Int, val movieName:String, val movieId:Int)
+data class ActorMovies(val actorId: Int, val actorName: String, val movieRate: Double, val movieYear: Int, val movieName:String, val movieId:Int)
