+++ app/src/main/java/com/example/actorsdatabaseapp/sql/data/model/Movie.kt
@@ -1,4 +1,4 @@
 package com.example.actorsdatabaseapp.sql.data.model
 
-data class Movie(val id: Int = 0, val year: Int, val imdbRate: Int, val actorId: Int, val movieName:String) {
+data class Movie(val id: Int = 0, val year: Int, val imdbRate: Double, val actorId: Int, val movieName:String) {
 }
